Move bip39 tests
