+++ tests/bip39.rs
@@ -1,0 +1,466 @@
+// Copyright 2020 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+#![cfg(feature = "bip39")]
+
+extern crate alloc;
+
+use alloc::vec::Vec;
+use rand::{rngs::OsRng, RngCore};
+use unicode_normalization::UnicodeNormalization;
+
+use crypto::keys::bip39::*;
+
+struct TestVector {
+    wordlist: wordlist::Wordlist<'static>,
+    entropy: &'static str,
+    mnemonic: &'static str,
+    passphrase: &'static str,
+    seed: &'static str,
+}
+
+#[test]
+fn test_vectors() {
+    let tvs = [
+        // https://github.com/bip32JP/bip32JP.github.io/blob/master/test_JP_BIP39.json
+        // mangled using utils/test_vectors/bip39/bip32JP.sh
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "00000000000000000000000000000000",
+            mnemonic: "e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e3818ae3819ee38289",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "a262d6fb6122ecf45be09c50492b31f92e9beb7d9a845987a02cefda57a15f9c467a17872029a9e92299b5cbdf306e3a0ee620245cbd508959b6cb7ca637bd55",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f",
+            mnemonic: "e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38199e38080e3828ae3818fe381a4e38080e381b0e38184e3818be38080e3828de3819be38293e38080e38284e381a1e38293e38080e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38281",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "aee025cbe6ca256862f889e48110a6a382365142f7d16f2b9545285b3af64e542143a577e9c144e101a6bdca18f8d97ec3366ebf5b088b1c1af9bc31346e60d9",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "80808080808080808080808080808080",
+            mnemonic: "e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e38193e3818ce3828ce3828be38080e38184e3818fe381b6e38293e38080e38191e38184e38191e38293e38080e38182e3819fe38188e3828be38080e38184e38288e3818fe38080e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e3818be381a1e38283e38293",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "e51736736ebdf77eda23fa17e31475fa1d9509c78f1deb6b4aacfbd760a7e2ad769c714352c95143b5c1241985bcb407df36d64e75dd5a2b78ca5d2ba82a3544",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "ffffffffffffffffffffffffffffffff",
+            mnemonic: "e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828de38293e381b6e38293",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "4cd2ef49b479af5e1efbbd1e0bdc117f6a29b1010211df4f78e2ed40082865793e57949236c43b9fe591ec70e5bb4298b8b71dc4b267bb96ed4ed282c8f7761c",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "000000000000000000000000000000000000000000000000",
+            mnemonic: "e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38289e38184e38190e381be",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "d99e8f1ce2d4288d30b9c815ae981edd923c01aa4ffdc5dee1ab5fe0d4a3e13966023324d119105aff266dac32e5cd11431eeca23bbd7202ff423f30d6776d69",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f",
+            mnemonic: "e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38199e38080e3828ae3818fe381a4e38080e381b0e38184e3818be38080e3828de3819be38293e38080e38284e381a1e38293e38080e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38199e38080e3828ae3818fe381a4e38080e381b0e38184e3818be38080e3828de3819be38293e38080e38284e381a1e38293e38080e3819de381a4e38186e38080e3828ce38184e3818e",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "eaaf171efa5de4838c758a93d6c86d2677d4ccda4a064a7136344e975f91fe61340ec8a615464b461d67baaf12b62ab5e742f944c7bd4ab6c341fbafba435716",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "808080808080808080808080808080808080808080808080",
+            mnemonic: "e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e38193e3818ce3828ce3828be38080e38184e3818fe381b6e38293e38080e38191e38184e38191e38293e38080e38182e3819fe38188e3828be38080e38184e38288e3818fe38080e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e38193e3818ce3828ce3828be38080e38184e3818fe381b6e38293e38080e38191e38184e38191e38293e38080e38182e3819fe38188e3828be38080e38184e38288e3818fe38080e3819de381a8e381a5e38289e38080e38184e3818de381aae3828a",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "aec0f8d3167a10683374c222e6e632f2940c0826587ea0a73ac5d0493b6a632590179a6538287641a9fc9df8e6f24e01bf1be548e1f74fd7407ccd72ecebe425",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "ffffffffffffffffffffffffffffffffffffffffffffffff",
+            mnemonic: "e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828ae38293e38194",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "f0f738128a65b8d1854d68de50ed97ac1831fc3a978c569e415bbcb431a6a671d4377e3b56abd518daa861676c4da75a19ccb41e00c37d086941e471a4374b95",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "0000000000000000000000000000000000000000000000000000000000000000",
+            mnemonic: "e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38182e38184e38193e3818fe38197e38293e38080e38184e381a3e381a6e38184",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "23f500eec4a563bf90cfda87b3e590b211b959985c555d17e88f46f7183590cd5793458b094a4dccc8f05807ec7bd2d19ce269e20568936a751f6f1ec7c14ddd",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f",
+            mnemonic: "e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38199e38080e3828ae3818fe381a4e38080e381b0e38184e3818be38080e3828de3819be38293e38080e38284e381a1e38293e38080e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38199e38080e3828ae3818fe381a4e38080e381b0e38184e3818be38080e3828de3819be38293e38080e38284e381a1e38293e38080e3819de381a4e38186e38080e3828ce3818de381a0e38184e38080e381bbe38293e38284e3818fe38080e3828fe3818be38199e38080e3828ae3818fe381a4e38080e381b0e38184e3818be38080e3828de3819be38293e38080e381bee38293e3818de381a4",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "cd354a40aa2e241e8f306b3b752781b70dfd1c69190e510bc1297a9c5738e833bcdc179e81707d57263fb7564466f73d30bf979725ff783fb3eb4baa86560b05",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "8080808080808080808080808080808080808080808080808080808080808080",
+            mnemonic: "e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e38193e3818ce3828ce3828be38080e38184e3818fe381b6e38293e38080e38191e38184e38191e38293e38080e38182e3819fe38188e3828be38080e38184e38288e3818fe38080e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e38193e3818ce3828ce3828be38080e38184e3818fe381b6e38293e38080e38191e38184e38191e38293e38080e38182e3819fe38188e3828be38080e38184e38288e3818fe38080e3819de381a8e381a5e38289e38080e38182e381bee381a9e38080e3818ae3818ae38186e38080e38182e38193e3818ce3828ce3828be38080e38184e3818fe381b6e38293e38080e38191e38184e38191e38293e38080e38182e3819fe38188e3828be38080e38186e38281e3828b",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "6b7cd1b2cdfeeef8615077cadd6a0625f417f287652991c80206dbd82db17bf317d5c50a80bd9edd836b39daa1b6973359944c46d3fcc0129198dc7dc5cd0e68",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+            mnemonic: "e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e3828fe3828ce3828be38080e38289e38184e38186",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "a44ba7054ac2f9226929d56505a51e13acdaa8a9097923ca07ea465c4c7e294c038f3f4e7e4b373726ba0057191aced6e48ac8d183f3a11569c426f0de414623",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "77c2b00716cec7213839159e404db50d",
+            mnemonic: "e3819be381bee38184e38080e38186e381a1e3818ce3828fe38080e38182e3819ae3818de38080e3818be3828de38186e38080e38281e3819ae38289e38197e38184e38080e381a0e38293e381a1e38080e381bee38199e3818fe38080e3818ae38195e38281e3828be38080e381a6e38184e381bce38186e38080e38182e3819fe3828be38080e38199e38182e381aae38080e38188e38197e38283e3818f",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "344cef9efc37d0cb36d89def03d09144dd51167923487eec42c487f7428908546fa31a3c26b7391a2b3afe7db81b9f8c5007336b58e269ea0bd10749a87e0193",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "b63a9c59a6e641f288ebc103017f1da9f8290b3da6bdef7b",
+            mnemonic: "e381ace38199e38280e38080e381b5e381a3e3818be381a4e38080e38186e381a9e38293e38080e38193e38186e3828ae381a4e38080e38197e381a4e38198e38080e3828ae38287e38186e3828ae38080e3818ae3819fe3818ce38184e38080e3819be38282e3819fe3828ce38080e38182e381a4e38281e3828be38080e38184e381a1e3828ae38285e38186e38080e381afe38293e38197e38283e38080e38194e381bee38199e3828ae38080e3819de38293e38191e38184e38080e3819fe38184e381a1e38287e38186e38080e38289e38197e38293e381b0e38293e38080e381b6e38293e3819be3818de38080e38284e38199e381bfe38080e381bbe38184e3818f",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "b14e7d35904cb8569af0d6a016cee7066335a21c1c67891b01b83033cadb3e8a034a726e3909139ecd8b2eb9e9b05245684558f329b38480e262c1d6bc20ecc4",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "3e141609b97933b66a060dcddc71fad1d91677db872031e85f4c015c5e7e8982",
+            mnemonic: "e3818fe381aee38186e38080e381a6e381ace38190e38184e38080e3819de38293e3818be38184e38080e38199e3828de381a3e381a8e38080e381a1e3818de38285e38186e38080e381bbe38182e38293e38080e381a8e38195e3818be38080e381afe3818fe38197e38285e38080e381b2e381b3e3818fe38080e381bfe38188e3828be38080e3819de38196e38184e38080e381a6e38293e38199e38186e38080e3819fe38293e381b4e38293e38080e3818fe38197e38287e38186e38080e38199e38184e38288e38186e381b3e38080e381bfe38191e38293e38080e3818de38195e38289e3818ee38080e38192e38196e38293e38080e381b5e3818fe38196e381a4e38080e38182e381a4e3818be38186e38080e381afe38284e38184e38080e3818fe3828de38186e38080e3818ae38284e38286e381b3e38080e38193e38199e38186",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "32e78dce2aff5db25aa7a4a32b493b5d10b4089923f3320c8b287a77e512455443298351beb3f7eb2390c4662a2e566eec5217e1a37467af43b46668d515e41b",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "0460ef47585604c5660618db2e6a7e7f",
+            mnemonic: "e38182e381bfe38282e381aee38080e38184e3818de3818ae38184e38080e381b5e38184e38186e381a1e38080e381abe38192e3828be38080e38196e38293e38197e38287e38080e38198e3818be38293e38080e381a4e38184e3818be38080e381afe3819fe38293e38080e381bbe38182e38293e38080e38199e38293e381bde38186e38080e381a6e381a1e3818ce38184e38080e3828fe3818be38281",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "0acf902cd391e30f3f5cb0605d72a4c849342f62bd6a360298c7013d714d7e58ddf9c7fdf141d0949f17a2c9c37ced1d8cb2edabab97c4199b142c829850154b",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "72f60ebac5dd8add8d2a25a797102c3ce21bc029c200076f",
+            mnemonic: "e38199e3828de381a3e381a8e38080e381abe3818fe38197e381bfe38080e381aae38284e38280e38080e3819fe381a8e38188e3828be38080e381b8e38184e38193e38186e38080e38199e3818fe38186e38080e3818de381aae38184e38080e38191e381a3e381a6e38184e38080e381a8e3818fe381b9e381a4e38080e381ade381a3e38197e38293e38080e38184e3819fe381bfe38080e3819be38293e3819be38184e38080e3818ae3818fe3828ae3818ce381aae38080e381bee3818be38184e38080e381a8e3818fe38184e38080e38191e38182e381aae38080e38184e3818de3818ae38184e38080e3819de3819de38190",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "9869e220bec09b6f0c0011f46e1f9032b269f096344028f5006a6e69ea5b0b8afabbb6944a23e11ebd021f182dd056d96e4e3657df241ca40babda532d364f73",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "2c85efc7f24ee4573d2b81a6ec66cee209b2dcbd09d8eddc51e0215b0b68e416",
+            mnemonic: "e3818be381bbe38194e38080e3818de38186e38184e38080e38286e3819fe3818be38080e381bfe38199e38188e3828be38080e38282e38289e38186e38080e3818ce381a3e38193e38186e38080e38288e3819de38186e38080e3819ae381a3e381a8e38080e381a8e3818de381a9e3818de38080e38197e3819fe38186e38191e38080e381abe38293e3818be38080e381afe381a3e38193e38186e38080e381a4e381bfe3818de38080e38199e38186e38198e381a4e38080e38288e38191e38184e38080e3818fe38192e38293e38080e38282e3818fe381a6e3818de38080e381bee3828fe3828ae38080e3819be38281e3828be38080e38192e38196e38184e38080e381abe38192e3828be38080e381abe38293e3819fe38184e38080e3819fe38293e3819de3818fe38080e381bbe3819de3818f",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "713b7e70c9fbc18c831bfd1f03302422822c3727a93a5efb9659bec6ad8d6f2c1b5c8ed8b0b77775feaf606e9d1cc0a84ac416a85514ad59f5541ff5e0382481",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "eaebabb2383351fd31d703840b32e9e2",
+            mnemonic: "e38281e38184e38188e38293e38080e38195e381aee38186e38080e38281e381a0e381a4e38080e38199e381a6e3828be38080e3818de381ace38194e38197e38080e3828de38293e381b1e38080e381afe38293e38193e38080e381bee38191e3828be38080e3819fe38184e3818ae38186e38080e38195e3818be38184e38197e38080e381ade38293e38184e3828ae38080e381afe381b6e38289e38197",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "06e1d5289a97bcc95cb4a6360719131a786aba057d8efd603a547bd254261c2a97fcd3e8a4e766d5416437e956b388336d36c7ad2dba4ee6796f0249b10ee961",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "7ac45cfe7722ee6c7ba84fbc2d5bd61b45cb2fe5eb65aa78",
+            mnemonic: "e3819be38293e381b1e38184e38080e3818ae38197e38188e3828be38080e38190e38293e3818be38293e38080e38282e38289e38186e38080e3818de38182e38184e38080e3818de381bce38186e38080e38284e3818ae38284e38080e38184e3819be38188e381b3e38080e381aee38184e3819ae38080e38198e38285e38197e38293e38080e38288e38286e38186e38080e3818de381bfe381a4e38080e38195e381a8e38184e38282e38080e381a1e38293e38282e3818fe38080e381a1e3828fe3828fe38080e38197e38293e3819be38184e38198e38080e381a8e38281e3828be38080e381afe381a1e381bfe381a4",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "1fef28785d08cbf41d7a20a3a6891043395779ed74503a5652760ee8c24dfe60972105ee71d5168071a35ab7b5bd2f8831f75488078a90f0926c8e9171b2bc4a",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "4fa1a8bc3e6d80ee1316050e862c1812031493212b7ec3f3bb1b08f168cabeef",
+            mnemonic: "e38193e38193e3828de38080e38184e381a9e38186e38080e3818de38182e381a4e38080e3819de38186e3818ce38293e3818de38287e38186e38080e381b8e38184e38182e38293e38080e3819be381a4e3828ae381a4e38080e38194e38186e3819be38184e38080e381afe38184e381a1e38080e38184e381b3e3818de38080e3818de38193e3818fe38080e38182e38293e38184e38080e3818ae381a1e381a4e3818fe38080e3818de38193e38188e3828be38080e38191e38293e381a8e38186e38080e3819fe38184e38193e38080e38199e38199e38281e3828be38080e381afe381a3e38191e38293e38080e381a6e38184e381a9e38080e381afe38293e3818ae38293e38080e38184e38293e38195e381a4e38080e38186e381aae3818ee38080e38197e381ade381bee38080e3828ce38184e381bce38186e38080e381bfe381a4e3818be3828b",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "43de99b502e152d4c198542624511db3007c8f8f126a30818e856b2d8a20400d29e7a7e3fdd21f909e23be5e3c8d9aee3a739b0b65041ff0b8637276703f65c2",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "18ab19a9f54a9274f03e5209a2ac8a91",
+            mnemonic: "e38186e3828ae3818de3828ce38080e38195e38184e3819be38184e38080e38198e38286e38186e38080e38280e3828de38293e38080e381a8e381a9e38191e3828be38080e38190e38186e3819fe38289e38080e381afe38184e3828ce381a4e38080e381b2e38191e381a4e38080e38184e3819ae3828ce38080e38186e381a1e38182e3828fe3819be38080e3818ae38195e38281e3828be38080e3818ae3819fe3818f",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "3d711f075ee44d8b535bb4561ad76d7d5350ea0b1f5d2eac054e869ff7963cdce9581097a477d697a2a9433a0c6884bea10a2193647677977c9820dd0921cbde",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "18a2e1d81b8ecfb2a333adcb0c17a5b9eb76cc5d05db91a4",
+            mnemonic: "e38186e3828ae3818de3828ce38080e38186e381ade3828be38080e3819be381a3e38195e3819fe3818fe381bee38080e3818de38282e381a1e38080e38281e38293e3818de38287e38080e381b8e38184e3819fe3818fe38080e3819fe381bee38194e38080e3819ce381a3e3818fe38080e381b3e38198e38285e381a4e3818be38293e38080e38195e38293e3819de38080e38280e3819be3828be38080e3819be38184e38198e38080e381ade3818fe3819fe38184e38080e38197e381afe38289e38184e38080e3819be3818ae38186e38080e381ade38293e381a9e38080e3819fe38293e381bee381a4e38080e3818ce38184e38191e38293",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "753ec9e333e616e9471482b4b70a18d413241f1e335c65cd7996f32b66cf95546612c51dcf12ead6f805f9ee3d965846b894ae99b24204954be80810d292fcdd",
+        },
+        TestVector {
+            wordlist: wordlist::JAPANESE,
+            entropy: "15da872c95a13dd738fbf50e427583ad61f18fd99f628c417a61cf8343c90419",
+            mnemonic: "e38186e381a1e38285e38186e38080e381b5e3819de3818fe38080e381b2e38197e38287e38080e3818ce381a1e38287e38186e38080e38186e38191e38282e381a4e38080e38281e38184e3819de38186e38080e381bfe3818be38293e38080e3819de38196e38184e38080e38184e381b0e3828be38080e38186e38191e381a8e3828be38080e38195e38293e381bee38080e38195e38193e381a4e38080e3818ae38186e38195e381bee38080e381b1e38293e381a4e38080e38197e381b2e38287e38186e38080e38281e38197e3819fe38080e3819fe381afe381a4e38080e38184e381a1e381b6e38080e381a4e38186e38198e38287e38186e38080e381a6e38195e3818ee38287e38186e38080e3818de381a4e381ade38080e381bfe38199e38188e3828be38080e38184e3828ae38190e381a1e38080e3818be38281e3828ce3818ae38293",
+            passphrase: "e38d8de382ace38390e383b4e382a1e381b1e381b0e38190e3829ee381a1e381a2e58d81e4babae58d81e889b2",
+            seed: "346b7321d8c04f6f37b49fdf062a2fddc8e1bf8f1d33171b65074531ec546d1d3469974beccb1a09263440fc92e1042580a557fdce314e27ee4eabb25fa5e5fe",
+        },
+
+        // https://github.com/trezor/python-mnemonic/blob/master/vectors.json
+        // mangled using utils/test_vectors/bip39/trezor.sh
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "00000000000000000000000000000000",
+            mnemonic: "6162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e2061626f7574",
+            passphrase: "5452455a4f52",
+            seed: "c55257c360c07c72029aebc1b53c05ed0362ada38ead3e3e9efa3708e53495531f09a6987599d18264c1e1c92f2cf141630c7a3c4ab7c81b2f001698e7463b04",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f",
+            mnemonic: "6c6567616c2077696e6e6572207468616e6b20796561722077617665207361757361676520776f7274682075736566756c206c6567616c2077696e6e6572207468616e6b2079656c6c6f77",
+            passphrase: "5452455a4f52",
+            seed: "2e8905819b8723fe2c1d161860e5ee1830318dbf49a83bd451cfb8440c28bd6fa457fe1296106559a3c80937a1c1069be3a3a5bd381ee6260e8d9739fce1f607",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "80808080808080808080808080808080",
+            mnemonic: "6c65747465722061647669636520636167652061627375726420616d6f756e7420646f63746f722061636f75737469632061766f6964206c65747465722061647669636520636167652061626f7665",
+            passphrase: "5452455a4f52",
+            seed: "d71de856f81a8acc65e6fc851a38d4d7ec216fd0796d0a6827a3ad6ed5511a30fa280f12eb2e47ed2ac03b5c462a0358d18d69fe4f985ec81778c1b370b652a8",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "ffffffffffffffffffffffffffffffff",
+            mnemonic: "7a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f2077726f6e67",
+            passphrase: "5452455a4f52",
+            seed: "ac27495480225222079d7be181583751e86f571027b0497b5b5d11218e0a8a13332572917f0f8e5a589620c6f15b11c61dee327651a14c34e18231052e48c069",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "000000000000000000000000000000000000000000000000",
+            mnemonic: "6162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206167656e74",
+            passphrase: "5452455a4f52",
+            seed: "035895f2f481b1b0f01fcf8c289c794660b289981a78f8106447707fdd9666ca06da5a9a565181599b79f53b844d8a71dd9f439c52a3d7b3e8a79c906ac845fa",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f",
+            mnemonic: "6c6567616c2077696e6e6572207468616e6b20796561722077617665207361757361676520776f7274682075736566756c206c6567616c2077696e6e6572207468616e6b20796561722077617665207361757361676520776f7274682075736566756c206c6567616c2077696c6c",
+            passphrase: "5452455a4f52",
+            seed: "f2b94508732bcbacbcc020faefecfc89feafa6649a5491b8c952cede496c214a0c7b3c392d168748f2d4a612bada0753b52a1c7ac53c1e93abd5c6320b9e95dd",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "808080808080808080808080808080808080808080808080",
+            mnemonic: "6c65747465722061647669636520636167652061627375726420616d6f756e7420646f63746f722061636f75737469632061766f6964206c65747465722061647669636520636167652061627375726420616d6f756e7420646f63746f722061636f75737469632061766f6964206c657474657220616c77617973",
+            passphrase: "5452455a4f52",
+            seed: "107d7c02a5aa6f38c58083ff74f04c607c2d2c0ecc55501dadd72d025b751bc27fe913ffb796f841c49b1d33b610cf0e91d3aa239027f5e99fe4ce9e5088cd65",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "ffffffffffffffffffffffffffffffffffffffffffffffff",
+            mnemonic: "7a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207768656e",
+            passphrase: "5452455a4f52",
+            seed: "0cd6e5d827bb62eb8fc1e262254223817fd068a74b5b449cc2f667c3f1f985a76379b43348d952e2265b4cd129090758b3e3c2c49103b5051aac2eaeb890a528",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "0000000000000000000000000000000000000000000000000000000000000000",
+            mnemonic: "6162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e206162616e646f6e20617274",
+            passphrase: "5452455a4f52",
+            seed: "bda85446c68413707090a52022edd26a1c9462295029f2e60cd7c4f2bbd3097170af7a4d73245cafa9c3cca8d561a7c3de6f5d4a10be8ed2a5e608d68f92fcc8",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f7f",
+            mnemonic: "6c6567616c2077696e6e6572207468616e6b20796561722077617665207361757361676520776f7274682075736566756c206c6567616c2077696e6e6572207468616e6b20796561722077617665207361757361676520776f7274682075736566756c206c6567616c2077696e6e6572207468616e6b20796561722077617665207361757361676520776f727468207469746c65",
+            passphrase: "5452455a4f52",
+            seed: "bc09fca1804f7e69da93c2f2028eb238c227f2e9dda30cd63699232578480a4021b146ad717fbb7e451ce9eb835f43620bf5c514db0f8add49f5d121449d3e87",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "8080808080808080808080808080808080808080808080808080808080808080",
+            mnemonic: "6c65747465722061647669636520636167652061627375726420616d6f756e7420646f63746f722061636f75737469632061766f6964206c65747465722061647669636520636167652061627375726420616d6f756e7420646f63746f722061636f75737469632061766f6964206c65747465722061647669636520636167652061627375726420616d6f756e7420646f63746f722061636f757374696320626c657373",
+            passphrase: "5452455a4f52",
+            seed: "c0c519bd0e91a2ed54357d9d1ebef6f5af218a153624cf4f2da911a0ed8f7a09e2ef61af0aca007096df430022f7a2b6fb91661a9589097069720d015e4e982f",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+            mnemonic: "7a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f207a6f6f20766f7465",
+            passphrase: "5452455a4f52",
+            seed: "dd48c104698c30cfe2b6142103248622fb7bb0ff692eebb00089b32d22484e1613912f0a5b694407be899ffd31ed3992c456cdf60f5d4564b8ba3f05a69890ad",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "9e885d952ad362caeb4efe34a8e91bd2",
+            mnemonic: "6f7a6f6e65206472696c6c2067726162206669626572206375727461696e2067726163652070756464696e67207468616e6b2063727569736520656c646572206569676874207069636e6963",
+            passphrase: "5452455a4f52",
+            seed: "274ddc525802f7c828d8ef7ddbcdc5304e87ac3535913611fbbfa986d0c9e5476c91689f9c8a54fd55bd38606aa6a8595ad213d4c9c9f9aca3fb217069a41028",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "6610b25967cdcca9d59875f5cb50b0ea75433311869e930b",
+            mnemonic: "67726176697479206d616368696e65206e6f72746820736f72742073797374656d2066656d616c652066696c74657220617474697475646520766f6c756d6520666f6c6420636c756220737461792066656174757265206f66666963652065636f6c6f677920737461626c65206e6172726f7720666f67",
+            passphrase: "5452455a4f52",
+            seed: "628c3827a8823298ee685db84f55caa34b5cc195a778e52d45f59bcf75aba68e4d7590e101dc414bc1bbd5737666fbbef35d1f1903953b66624f910feef245ac",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "68a79eaca2324873eacc50cb9c6eca8cc68ea5d936f98787c60c7ebc74e6ce7c",
+            mnemonic: "68616d73746572206469616772616d20707269766174652064757463682063617573652064656c61792070726976617465206d65617420736c69646520746f64646c65722072617a6f7220626f6f6b2068617070792066616e637920676f7370656c2074656e6e6973206d61706c652064696c656d6d61206c6f616e20776f726420736872756720696e666c6963742064656c6179206c656e677468",
+            passphrase: "5452455a4f52",
+            seed: "64c87cde7e12ecf6704ab95bb1408bef047c22db4cc7491c4271d170a1b213d20b385bc1588d9c7b38f1b39d415665b8a9030c9ec653d75e65f847d8fc1fc440",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "c0ba5a8e914111210f2bd131f3d5e08d",
+            mnemonic: "736368656d652073706f742070686f746f20636172642062616279206d6f756e7461696e20646576696365206b69636b20637261646c652070616374206a6f696e20626f72726f77",
+            passphrase: "5452455a4f52",
+            seed: "ea725895aaae8d4c1cf682c1bfd2d358d52ed9f0f0591131b559e2724bb234fca05aa9c02c57407e04ee9dc3b454aa63fbff483a8b11de949624b9f1831a9612",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "6d9be1ee6ebd27a258115aad99b7317b9c8d28b6d76431c3",
+            mnemonic: "686f726e2074656e616e74206b6e65652074616c656e742073706f6e736f72207370656c6c206761746520636c69702070756c736520736f617020736c757368207761726d2073696c766572206e6570686577207377617020756e636c6520637261636b206272617665",
+            passphrase: "5452455a4f52",
+            seed: "fd579828af3da1d32544ce4db5c73d53fc8acc4ddb1e3b251a31179cdb71e853c56d2fcb11aed39898ce6c34b10b5382772db8796e52837b54468aeb312cfc3d",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "9f6a2878b2520799a44ef18bc7df394e7061a224d2c33cd015b157d746869863",
+            mnemonic: "70616e64612065796562726f772062756c6c657420676f72696c6c612063616c6c20736d6f6b65206d756666696e207461737465206d65736820646973636f76657220736f6674206f73747269636820616c636f686f6c207370656564206e6174696f6e20666c617368206465766f7465206c6576656c20686f62627920717569636b20696e6e65722064726976652067686f737420696e73696465",
+            passphrase: "5452455a4f52",
+            seed: "72be8e052fc4919d2adf28d5306b5474b0069df35b02303de8c1729c9538dbb6fc2d731d5f832193cd9fb6aeecbc469594a70e3dd50811b5067f3b88b28c3e8d",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "23db8160a31d3e0dca3688ed941adbf3",
+            mnemonic: "636174207377696e6720666c61672065636f6e6f6d79207374616469756d20616c6f6e6520636875726e20737065656420756e69717565207061746368207265706f727420747261696e",
+            passphrase: "5452455a4f52",
+            seed: "deb5f45449e615feff5640f2e49f933ff51895de3b4381832b3139941c57b59205a42480c52175b6efcffaa58a2503887c1e8b363a707256bdd2b587b46541f5",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "8197a4a47f0425faeaa69deebc05ca29c0a5b5cc76ceacc0",
+            mnemonic: "6c696768742072756c652063696e6e616d6f6e2077726170206472617374696320776f726420707269646520737175697272656c2075706772616465207468656e20696e636f6d6520666174616c206170617274207375737461696e20637261636b20737570706c792070726f756420616363657373",
+            passphrase: "5452455a4f52",
+            seed: "4cbdff1ca2db800fd61cae72a57475fdc6bab03e441fd63f96dabd1f183ef5b782925f00105f318309a7e9c3ea6967c7801e46c8a58082674c860a37b93eda02",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "066dca1a2bb7e8a1db2832148ce9933eea0f3ac9548d793112d9a95c9407efad",
+            mnemonic: "616c6c20686f7572206d616b65206669727374206c656164657220657874656e6420686f6c6520616c69656e20626568696e6420677561726420676f7370656c206c6176612070617468206f75747075742063656e737573206d757365756d206a756e696f72206d6173732072656f70656e2066616d6f75732073696e6720616476616e63652073616c74207265666f726d",
+            passphrase: "5452455a4f52",
+            seed: "26e975ec644423f4a4c4f4215ef09b4bd7ef924e85d1d17c4cf3f136c2863cf6df0a475045652c57eb5fb41513ca2a2d67722b77e954b4b3fc11f7590449191d",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "f30f8c1da665478f49b001d94c5fc452",
+            mnemonic: "76657373656c206c616464657220616c746572206572726f72206665646572616c207369626c696e672063686174206162696c6974792073756e20676c6173732076616c76652070696374757265",
+            passphrase: "5452455a4f52",
+            seed: "2aaa9242daafcee6aa9d7269f17d4efe271e1b9a529178d7dc139cd18747090bf9d60295d0ce74309a78852a9caadf0af48aae1c6253839624076224374bc63f",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "c10ec20dc3cd9f652c7fac2f1230f7a3c828389a14392f05",
+            mnemonic: "73636973736f727320696e76697465206c6f636b206d61706c652073757072656d652072617720726170696420766f696420636f6e6772657373206d7573636c65206469676974616c20656c6567616e74206c6974746c6520627269736b2068616972206d616e676f20636f6e677265737320636c756d70",
+            passphrase: "5452455a4f52",
+            seed: "7b4a10be9d98e6cba265566db7f136718e1398c71cb581e1b2f464cac1ceedf4f3e274dc270003c670ad8d02c4558b2f8e39edea2775c9e232c7cb798b069e88",
+        },
+        TestVector {
+            wordlist: wordlist::ENGLISH,
+            entropy: "f585c11aec520db57dd353c69554b21a89b20fb0650966fa0a9d6f74fd989d8f",
+            mnemonic: "766f696420636f6d65206566666f7274207375666665722063616d70207375727665792077617272696f722068656176792073686f6f74207072696d61727920636c75746368206372757368206f70656e20616d617a696e672073637265656e20706174726f6c2067726f757020737061636520706f696e742074656e20657869737420736c75736820696e766f6c766520756e666f6c64",
+            passphrase: "5452455a4f52",
+            seed: "01f5bced59dec48e362f2c45b5de68b9fd6c92c6634f44d6d40aab69056506f0e35524a518034ddc1192e1dacd32c1ed3eaa3c3b131c88ed8e7e54c49a5d0998",
+        },
+    ];
+
+    for tv in tvs.iter() {
+        let entropy = hex::decode(tv.entropy).unwrap();
+        let mnemonic = hex::decode(tv.mnemonic).unwrap();
+        let mnemonic = core::str::from_utf8(&mnemonic).unwrap();
+
+        assert_eq!(
+            wordlist::encode(&entropy, &tv.wordlist)
+                .unwrap()
+                .nfkd()
+                .collect::<String>(),
+            mnemonic.nfkd().collect::<String>()
+        );
+
+        assert_eq!(wordlist::decode(mnemonic, &tv.wordlist).unwrap(), entropy);
+
+        let passphrase = hex::decode(tv.passphrase).unwrap();
+        let passphrase = core::str::from_utf8(&passphrase).unwrap();
+        let mut expected_seed = [0; 64];
+        hex::decode_to_slice(tv.seed, &mut expected_seed).unwrap();
+
+        let mut seed = [0; 64];
+        mnemonic_to_seed(mnemonic, passphrase, &mut seed);
+        assert_eq!(seed, expected_seed);
+    }
+}
+
+const ALL_WORDLISTS: &[wordlist::Wordlist<'static>] = &[wordlist::ENGLISH, wordlist::JAPANESE];
+fn choose_wordlist() -> &'static wordlist::Wordlist<'static> {
+    &ALL_WORDLISTS[rand::random::<usize>() % ALL_WORDLISTS.len()]
+}
+
+#[test]
+fn test_wordlist_codec() {
+    for _ in 0..1000 {
+        let mut data = vec![0; 32 * (4 + rand::random::<usize>() % 5) / 8];
+        OsRng.fill_bytes(&mut data);
+
+        let ws = choose_wordlist();
+
+        let ms = wordlist::encode(&data, ws).unwrap();
+        assert_eq!(wordlist::decode(&ms, ws).unwrap(), data);
+        assert_eq!(wordlist::verify(&ms, ws), Ok(()));
+    }
+}
+
+#[test]
+fn test_wordlist_codec_different_data_different_encodings() {
+    for _ in 0..1000 {
+        let mut data = vec![0; 32 * (4 + rand::random::<usize>() % 5) / 8];
+        OsRng.fill_bytes(&mut data);
+
+        let mut corrupted_data = data.clone();
+        crate::utils::test_utils::corrupt(&mut corrupted_data);
+
+        let ws = choose_wordlist();
+        let ms = wordlist::encode(&data, &ws).unwrap();
+
+        assert_ne!(ms, wordlist::encode(&corrupted_data, ws).unwrap());
+    }
+}
+
+#[test]
+#[allow(non_snake_case)]
+fn test_wordlist_codec_error_detection() {
+    for ENT in &[128, 160, 192, 224, 256] {
+        let mut false_positives = 0;
+        let CS = ENT / 32;
+        let N = 1000;
+        let acceptable_false_positives = 2 * N / (1 << CS);
+        for _ in 0..N {
+            let mut data = vec![0; ENT / 8];
+            OsRng.fill_bytes(&mut data);
+
+            let ws = choose_wordlist();
+            let ms = wordlist::encode(&data, ws).unwrap();
+
+            let mut wrong_word = ms.clone();
+            while wrong_word == ms {
+                wrong_word = ms
+                    .split(ws.separator)
+                    .map(|w| {
+                        if rand::random::<usize>() % 8 == 0 {
+                            ws.words[rand::random::<usize>() % 2048].to_string()
+                        } else {
+                            w.to_string()
+                        }
+                    })
+                    .collect::<Vec<String>>()
+                    .join(ws.separator);
+            }
+
+            if wordlist::decode(&wrong_word, ws) != Err(wordlist::Error::ChecksumMismatch) {
+                false_positives += 1;
+            }
+        }
+
+        assert!(false_positives <= acceptable_false_positives);
+    }
+}
